no code changes, just name changes and comments
